perplexity-alpaca-integration/requirements.txt
requests>=2.31.0
alpaca-py>=0.21.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pytest>=7.4.0
aiohttp>=3.8.0
websockets>=11.0
asyncio-mqtt>=0.16.0
pydantic>=2.0.0
loguru>=0.7.0
schedule>=1.2.0